add directory.build.props and stylecop.analyzers
