--- analyzers/Stylecop.Analyzers/Directory.Packages.props
@@ -1,11 +1,0 @@
-<Project>
-
-  <PropertyGroup>
-    <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageVersion Include="StyleCop.Analyzers.Unstable" Version="1.2.0.507" />
-  </ItemGroup>
-
-</Project>
