+++ dependency-injection/DependencyInjection/Properties/launchSettings.json
