+++ dependency-injection/DependencyInjection/Singleton.cs
