+++ source-generator/CodeGen/AnalyzerConfig.cs
@@ -2,8 +2,8 @@
 
 public class AnalyzerConfig
 {
-    public string ControllerServicesNamespace { get; set; }
-    public string JsonSchema { get; set; }
-    public string TargetProject { get; set; }
+    public string? ControllerServicesNamespace { get; set; }
+    public string? JsonSchema { get; set; }
+    public string? TargetProject { get; set; }
 }
