+++ source-generator/CodeGen/CodeGen.csproj
@@ -9,9 +9,9 @@
   </PropertyGroup>
 
   <ItemGroup>
-  	<PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.7.0" PrivateAssets="all" />
-  	<PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.3.4" PrivateAssets="all" />
-  	<PackageReference Include="Newtonsoft.Json" Version="13.0.3" GeneratePathProperty="true" PrivateAssets="all" />
+  	<PackageReference Include="Microsoft.CodeAnalysis.CSharp" PrivateAssets="all" />
+  	<PackageReference Include="Microsoft.CodeAnalysis.Analyzers" PrivateAssets="all" />
+  	<PackageReference Include="Newtonsoft.Json" GeneratePathProperty="true" PrivateAssets="all" />
   </ItemGroup>
 
   <PropertyGroup>
