+++ source-generator/CodeGen/Operation.cs
@@ -2,7 +2,7 @@
 
 public class Operation
 {
-    public string Name { get; set; }
-    public string Type { get; set; }
-    public object ReturnValue { get; set; }
+    public string? Name { get; set; }
+    public string? Type { get; set; }
+    public object? ReturnValue { get; set; }
 }
