+++ source-generator/CodeGen/ServiceModel.cs
@@ -2,8 +2,8 @@
 
 public class ServiceModel
 {
-    public string Namespace { get; set; }
-    public string TargetNamespace { get; set; }
-    public string Name { get; set; }
-    public Operation[] Operations { get; set; }
+    public string? Namespace { get; set; }
+    public string? TargetNamespace { get; set; }
+    public string? Name { get; set; }
+    public Operation[]? Operations { get; set; }
 }
