--- unit-testing/test/UnitTesting/Usings.cs
@@ -1,3 +1,0 @@
-global using NUnit.Framework;
-global using Moq;
-global using Shouldly;
